pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Kiroshi"
version = "0.0.0"
description = "Generic Alerting Platform"
authors = ["Bink DevOps <devops@bink.com>"]
readme = "README.md"

[tool.poetry.scripts]
kiroshi = "kiroshi.cli.cli:cli" # cli

[tool.poetry.dependencies]
python = "^3.11"
azure-identity = "^1.12.0"
azure-monitor-query = "^1.1.0"
sqlalchemy = "^2.0.2"
pydantic = "^1.10.4"
click = "^8.1.3"
alembic = "^1.9.3"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"
loguru = "^0.6.0"
pendulum = "^2.1.2"
beautifulsoup4 = "^4.11.2"
redis = "^4.5.1"
azure-keyvault = "^4.2.0"
trogon = "^0.2.1"
cryptography = "^41.0.0"
paramiko = "^3.2.0"
azure-storage-blob = "^12.16.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.275"

[tool.black]
line-length = 120
force-exclude = 'original_code/*'

[tool.ruff]
select = ["ALL"]
ignore = ["D203", "D213", "D407", "ANN101"]
line-length = 250
exclude = ["original_code/"]
